requirements.txt
aenum==2.2.6
attrs==23.2.0
behave==1.2.6
black==23.12.1
cassandra-driver==3.29.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.0
docker==7.0.0
geomet==0.2.1.post1
gremlinpython==3.4.6
idna==3.6
iniconfig==2.0.0
isodate==0.6.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==23.2
parse==1.20.0
parse-type==0.6.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pytest==7.4.4
python-snappy==0.6.1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.17.1
scales==1.0.9
six==1.16.0
tornado==5.1.1
]]]]]]]]]]]]]]]]]]